Complete day 23

+++ app/Http/Controllers/JobController.php
@@ -3,7 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Job;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class JobController extends Controller
 {
@@ -19,15 +22,18 @@ class JobController extends Controller
         ]);
     }
 
-    public function create() {
+    public function create() 
+    {
         return view('jobs.create');
     }
 
-    public function show(Job $job) {
+    public function show(Job $job) 
+    {
         return view('jobs.show', ['job' => $job]);
     }
 
-    public function store() {
+    public function store() 
+    {
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required'],
@@ -42,11 +48,15 @@ class JobController extends Controller
         return redirect('/jobs');
     }
 
-    public function edit(Job $job) {
+    public function edit(Job $job) 
+    {
         return view('jobs.edit', ['job' => $job]);
     }
 
-    public function update(Job $job) {
+    public function update(Job $job) 
+    {
+        Gate::authorize('edit-job', $job);
+
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required'],
@@ -60,8 +70,12 @@ class JobController extends Controller
         return redirect("/jobs/" . $job->id);
     }
 
-    public function destroy(Job $job) {
+    public function destroy(Job $job) 
+    {
+        Gate::authorize('edit-job', $job);
+
         $job->delete();
+
         return redirect("/jobs");
     }
 }
